requirements.txt
Flask==2.0.1
numpy==1.21.0
opencv-python==4.5.3
Pillow==8.3.1
rembg==2.0.30